mini_blog/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.1
psycopg2>=2.9.6  # Если используется PostgreSQL
mysqlclient>=2.1.1  # Если используется MySQL
gunicorn>=20.1.0  # Для деплоя